+++ charts/substra-backend/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.0.3
 
 ### Added
-- New channel configuration setting `enable_model_export`. When enabled, it allows authenticated users to download models trained on the current node.
+- Add new channel configuration setting `model_export_enabled`. When enabled, it allows authenticated users to download models trained on the current node.
 
 ## 2.0.2